src/requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.1
django-allauth==0.50.0
django-cleanup==6.0.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-filter==21.1
django-import-export==2.7.1
et-xmlfile==1.1.0
idna==3.3
kombu==5.2.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.26
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stripe==2.68.0
tablib==3.2.0
typing-extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
